Rename module, fix uspce test
