+++ src/telliot_feeds/feeds/spot_price_manual_feed.py
@@ -1,6 +1,6 @@
 from telliot_feeds.datafeed import DataFeed
 from telliot_feeds.queries.price.spot_price import SpotPrice
-from telliot_feeds.sources.manual_sources.spot_price_input_source import SpotPriceManualSource
+from telliot_feeds.sources.manual.spot_price_input_source import SpotPriceManualSource
 
 # Using defaults to bypass an error handle for feeds w/out non-manual sources
 asset: str = "eth"
