+++ src/telliot_feeds/feeds/string_query_feed.py
@@ -1,6 +1,6 @@
 from telliot_feeds.datafeed import DataFeed
 from telliot_feeds.queries.string_query import StringQuery
-from telliot_feeds.sources.manual_sources.string_query_manual_source import StringQueryManualSource
+from telliot_feeds.sources.manual.string_query_manual_source import StringQueryManualSource
 
 text = None
