+++ src/telliot_feeds/feeds/twap_manual_feed.py
@@ -2,7 +2,7 @@ import asyncio
 
 from telliot_feeds.datafeed import DataFeed
 from telliot_feeds.queries.price.twap import TWAP
-from telliot_feeds.sources.manual_sources.twap_manual_input_source import TWAPManualSource
+from telliot_feeds.sources.manual.twap_manual_input_source import TWAPManualSource
 
 
 # Using defaults to bypass an error handle for auto feeds and unsupported currencies
