+++ src/telliot_feeds/feeds/uspce_feed.py
@@ -1,7 +1,7 @@
 """Example datafeed used by USPCEReporter."""
 from telliot_feeds.datafeed import DataFeed
 from telliot_feeds.queries.ampleforth.uspce import AmpleforthUSPCE
-from telliot_feeds.sources.uspce import USPCESource
+from telliot_feeds.sources.manual.uspce import USPCESource
 
 
 uspce_feed = DataFeed(query=AmpleforthUSPCE(), source=USPCESource())
