+++ src/telliot_feeds/sources/manual/daily_volatility_manual_source.py
