+++ src/telliot_feeds/sources/manual/diva_manual_source.py
