+++ src/telliot_feeds/sources/manual/numeric_api_manual_response.py
