+++ src/telliot_feeds/sources/manual/spot_price_input_source.py
