+++ src/telliot_feeds/sources/manual/string_query_manual_source.py
